Split supervisor code into multiple files
