+++ src/supervisor/utils.h
@@ -1,0 +1,13 @@
+
+#ifndef R_PROCESSX_SUPERVISOR_UTILS_H
+#define R_PROCESSX_SUPERVISOR_UTILS_H
+
+#include <stdbool.h>
+
+
+extern bool verbose_mode;
+
+
+void verbose_printf(const char *format, ...);
+
+#endif
